requirements.txt
pyautogui
python-dotenv
requests
pillow 
pyscreeze
